Enabled memory test
