+++ test/Transactions/Orleans.Transactions.Tests/Memory/TransactionConcurrencyTests.cs
@@ -5,20 +5,11 @@ namespace Orleans.Transactions.Tests
 {
 
     [TestCategory("BVT"), TestCategory("Transactions")]
-    public class TransactionConcurrencyTests : TransactionConcurrencyTestRunner, IClassFixture<TransactionConcurrencyTests.SkipFixture>
+    public class TransactionConcurrencyTests : TransactionConcurrencyTestRunner, IClassFixture<MemoryTransactionsFixture>
     {
-        public TransactionConcurrencyTests(SkipFixture fixture, ITestOutputHelper output)
+        public TransactionConcurrencyTests(MemoryTransactionsFixture fixture, ITestOutputHelper output)
             : base(fixture.GrainFactory, output)
         {
-            fixture.EnsurePreconditionsMet();
-        }
-
-        public class SkipFixture : MemoryTransactionsFixture
-        {
-            protected override void CheckPreconditionsOrThrow()
-            {
-                throw new SkipException("Test fail against memory storage.  Remove SKipFixture when fixed");
-            }
         }
     }
 }
